tests

+++ src/pow.cpp
@@ -6,11 +6,11 @@
 class Pow {
 public:	
 	//функция итеративного возведения числа в степень	
-	double static simple_pow(double base, uint32_t exp) {
+	double static simple_pow(double base, uint64_t exp) {
 		
 		double result = 1; //присваиваем начальное значение - единицу (чтобы учесть возможное значение exp = 0)
 
-		for (uint32_t iter = 0; iter != exp; ++iter) {
+		for (uint64_t iter = 0; iter != exp; ++iter) {
 			result *= base;
 		}
 
@@ -18,7 +18,7 @@ public:
 	}
 	
 	//Функция возведения в степень через домножение
-	double static quick_pow(double base, uint32_t exp) {
+	double static quick_pow(double base, uint64_t exp) {
 		
 		if (0 == exp) return 1; //базовый случай
 
@@ -33,22 +33,22 @@ public:
 	}	
 
 	//Функция возведения в степень через двоичное разложение показателя степени
-	double static binary_pow(double base, uint32_t exp) {					
+	double static binary_pow(double base, uint64_t exp) {					
 		
 		double result = 1; //начальное значение
 
 		// Находим позицию старшего бита
-		int bit_pos = 0;
-		uint32_t temp = exp;
+		uint64_t bit_pos = 0;
+		uint64_t temp = exp;
 		while (temp > 1) {
 			temp >>= 1;
 			bit_pos++;
 		}
 
 		// Обрабатываем биты от старшего к младшему
-		for (int i = bit_pos; i >= 0; i--) {
+		for(int64_t i = bit_pos; i >= 0; i--) {
 			result *= result;
-			if (exp & (1 << i)) {
+			if (exp & (1ll << i)) {
 				result *= base;
 			}
 		}
